backend/api/requirements.txt
fastapi
gunicorn
uvicorn[standard]
pydantic
# llama-cpp-python installed separately in Dockerfile with CUDA support
redis
httpx
prometheus-client

# RAG Dependencies
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
rank_bm25>=0.2.2
nltk>=3.8.0
numpy>=1.24.0
